pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "qnbench"
version = "0.1.0"
description = "Quantum Network Routing Benchmark"
requires-python = ">=3.9"
dependencies = [
    "gymnasium>=0.29",
    "numpy>=1.24",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
rl = [
    "torch>=2.0",
]
dev = [
    "pytest>=7.0",
    "matplotlib>=3.7",
]
all = ["qnbench[rl,dev]"]

[tool.setuptools.packages.find]
include = ["qnbench*"]
